requirements.txt
blinker==1.9.0
click==8.1.7
defusedxml==0.7.1
Flask==3.1.0
fonttools==4.55.0
fpdf2==2.8.1
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
numpy==2.0.2
opencv-python==4.10.0.84
pillow==11.0.0
Werkzeug==3.1.3
zipp==3.21.0
